Fix navigation
Previous solution displaced the whole wrapper upward to hide the menu
leaving a blank space at the end of the page. It happened because the
displacement was caused by a CSS transformation which prevented the
browser to adapt to the content.

Current solution does not animate the whole wrapper. Instead, it
animates the menu and the content separately but in coordination so it
seems the menu pushes the content. Since now, the content is not
translated upwards but the other way around, it follows the scroll
direction avoiding the blank space.